requirements.txt
django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-dotenv==1.0.1
langchain-openai==0.2.11
chromadb==0.5.23
scikit-learn==1.5.2
pdfplumber==0.11.4
pandas==2.2.3
requests==2.32.3
streamlit==1.40.2
langchain==0.3.27